Fix release build
