+++ Meziantou.Xunit.ParallelTestFramework.Tests/ParallelMultiClassCollectionFixture.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace Meziantou.Xunit.ParallelTestFramework.Tests;
+
+[CollectionDefinition("ParallelMultiClass")]
+[EnableParallelization]
+public class ParallelMultiClassCollectionFixture : ICollectionFixture<CollectionConcurrencyFixture>
+{
+}
